requirements.txt
BeautifulSoup==3.2.1
EasyProcess==0.2.1
Flask==0.10.1
Flask-Cors==1.10.3
Jinja2==2.7.3
MarkupSafe==0.23
PyVirtualDisplay==0.1.5
PyYAML==3.11
Scrapy==1.0.3
# SPARQLWrapper==1.7.5
Werkzeug==0.10.1
YURL==0.13
argparse==1.2.1
beautifulsoup4==4.4.1
boto3==1.2.1
botocore==1.3.7
cffi==1.3.1
characteristic==14.3.0
cryptography==1.1.1
decorator==3.4.0
docutils==0.12
enum34==1.0.4
futures==2.2.0
html5lib==0.9999999
httplib2==0.9.2
idna==2.0
ipaddress==1.0.15
isodate==0.5.4
itsdangerous==0.24
jmespath==0.9.0
keepalive==0.5
lxml==3.5.0
microdata==0.5.3
ndg-httpsclient==0.4.0
networkx==1.9.1
nltk==3.1
numpy==1.10.1
-e git+https://github.com/Karmak23/opengraph.git@9006cbb4b40e5c0fca005b34dd1dc56ed00691d8#egg=opengraph-master
psycopg2==2.6.1
purl==1.1
pyOpenSSL==0.15.1
pyasn1==0.1.9
pyasn1-modules==0.0.8
pycparser==2.14
# -e git+https://github.com/RDFLib/pymicrodata.git@6d1088b11b21ec4b5fd0aeac8ff19b8de68e4f28#egg=pyMicrodata-master
python-dateutil==2.4.2
queuelib==1.4.2
# rdflib==4.2.1
requests==2.8.1
scikit-learn==0.17
scipy==0.16.1
selenium==2.48.0
service-identity==14.0.0
six==1.9.0
sklearn==0.0
wheel==0.24.0
wsgiref==0.1.2
